test: ERC721 behavior for Land
